+++ src/fetch.js
@@ -25,7 +25,7 @@ class Headers {
     if (init instanceof Headers || init instanceof Array)
       for (let [name, value] of init)
         this.append(name, value);
-    else if (typeof init === "object")
+    else if (typeof init === 'object')
       _.each(init, (value, name)=> {
         this.append(name, value);
       });